+++ HouseRules_Essentials/HouseRules_Essentials.csproj
@@ -48,6 +48,7 @@
     <Compile Include="Rulesets\NoSurprisesRuleset.cs" />
     <Compile Include="Rulesets\TheSwirlRuleset.cs" />
     <Compile Include="Rules\AbilityBackstabAdjustedRule.cs" />
+    <Compile Include="Rules\AbilityDamageOverriddenRule.cs" />
     <Compile Include="Rules\AbilityHealOverriddenRule.cs" />
     <Compile Include="Rules\CardClassRestrictionOverriddenRule.cs" />
     <Compile Include="Rules\TileEffectDurationOverriddenRule.cs" />
@@ -63,7 +64,6 @@
     <Compile Include="Rules\RegainAbilityIfMaxxedOutOverriddenRule.cs" />
     <Compile Include="Rules\StatModifiersOverridenRule.cs" />
     <Compile Include="Rules\AbilityRandomPieceListRule.cs" />
-    <Compile Include="Rules\AbilityDamageAdjustedRule.cs" />
     <Compile Include="Rules\CardAdditionOverriddenRule.cs" />
     <Compile Include="Rules\CardEnergyFromAttackMultipliedRule.cs" />
     <Compile Include="Rules\LevelExitLockedUntilAllEnemiesDefeatedRule.cs" />
